+++ IsekaiMod/Content/Features/IsekaiProtagonist/CharacterDevelopment/SneakyMagic.cs
@@ -1,5 +1,8 @@
 ﻿using IsekaiMod.Extensions;
 using IsekaiMod.Utilities;
+using Kingmaker.Blueprints;
+using Kingmaker.Blueprints.Classes;
+using Kingmaker.Blueprints.Classes.Prerequisites;
 using Kingmaker.Designers.Mechanics.Facts;
 using Kingmaker.UnitLogic.Abilities.Blueprints;
 using UnityEngine;
@@ -9,6 +12,7 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.CharacterDevelopment
     class SneakyMagic
     {
         private static readonly Sprite Icon_InvisibilityAlmostGreater = Resources.GetBlueprint<BlueprintAbility>("8dcb9c02148a704489948eaf84ab04bf").m_Icon;
+        private static readonly BlueprintFeature SneakAttack = Resources.GetBlueprint<BlueprintFeature>("9b9eac6709e1c084cb18c3a366e0ec87");
         public static void Add()
         {
             var SneakyMagic = Helpers.CreateFeature("SneakyMagic", bp => {
@@ -19,6 +23,9 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.CharacterDevelopment
                     + "the sneak attack damage.");
                 bp.m_Icon = Icon_InvisibilityAlmostGreater;
                 bp.AddComponent<SurpriseSpells>();
+                bp.AddComponent<PrerequisiteFeature>(c => {
+                    c.m_Feature = SneakAttack.ToReference<BlueprintFeatureReference>();
+                });
             });
             CharacterDevelopmentSelection.AddToSelection(SneakyMagic);
         }